implementing ExpenseClaims creation
- simplest possible implementation, needing a user_id and receipt_id,
which performs the "Submit for approval" users would do in the front-end
- made the call that Receipt should be able to be instantiated in XML
with minimal nodes (i.e. just the id) to allow referencing in
ExpenseClaims. In these cases, empty contact / line_item etc nodes are
pointless, so we now only add nodes that exist
- adjusting Receipt#contact to return nil if there is no receipt_from
specified, to avoid additional api calls when we only need a minimal
Receipt with an ID only
- removing the default date node from Receipt#to_xml - date should be
validated, so is no point adding || Date.today
- lacking validations and tests, naughty.